# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ENUMERA API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "name": "ENUMERA",
            "url": "https://enumera.tech",
            "email": "enumera@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/utilisateurs": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Liste des utilisateur",
                "description": "Retourne la liste des utilisateur",
                "operationId": "27a03cbccef26b92bf6d80712f8b8ab8",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Création d'un utilisateur",
                "description": "Crée une nouveau utilisateur et retourne le utilisateur créé",
                "operationId": "dded82ccae378b45b0c84f76ec078823",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "date_de_naissance",
                                    "genre",
                                    "telephone",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "date_de_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "profile": {
                                        "type": "string",
                                        "example": "ELEVE"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "75000000"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "@test@password#2024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "utilisateur créé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/utilisateurs/{slug}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Récupère un utilisateur par son slug",
                "description": "Retourne un utilisateur",
                "operationId": "8cd6984d6bae975d718d0045fa5f49e0",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du utilisateur à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Modification d'un utilisateur",
                "description": "Modifie un utilisateur et retourne le utilisateur modifié",
                "operationId": "b95b0d19daa0c0edea85d40137a0df6a",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du utilisateur à modifié",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "date_de_naissance",
                                    "genre",
                                    "telephone",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "date_de_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "profile": {
                                        "type": "string",
                                        "example": "ELEVE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "utilisateur modifié avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "utilisateur non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Suppression d'un utilisateur par son slug",
                "description": "Retourne le utilisateur supprimé",
                "operationId": "72955b9889fb84546f6b2a4d6f5f7040",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du utilisateur à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "utilisateur supprimé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "utilisateur non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}